extract/requirements.txt
requests
python-dotenv
dlt[postgres]>=1.5.0
dlt[duckdb]>=1.5.0
numpy
pandas
streamlit